pyproject.toml
[project]
name = "atlas-ops"
version = "0.4.2"
description = "Atlas Ops Copilot – automation toolkit with backend, agent, and web UI"
authors = [{name = "Atlas Ops Maintainers"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.3",
    "PyYAML>=6.0.1",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
]

[project.urls]
Homepage = "https://example.com/atlas-ops"

[project.scripts]
atlas-ops = "atlas_ops.cli:app"

[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["atlas_ops*"]

[tool.setuptools.package-data]
"atlas_ops" = ["frontend/*"]
